gammamellininvasymp: better fix for [#2283]
Converting to floating point speeds up considerably the function
(test-gchar-lfun goes down from 70s to 8s) but sometimes introduces
cancellation that eventually kills QD as in #2283.
Solution: start with exact input so that initial auxiliary data is exact
then convert to floating point (with 50% larger precision) later in
Klargeinit. This seems to work (and the original bad case is OK).