make gc_needed avoid GC if 'av' < (bot+top)/2 [#2330]
Rationale: if stack is already half spent when returning at the
reference point, GC is not that useful, increasing stack size is a
better bet.

As a result a few bench require a little more memory, no more than twice
the previous amount except in hypergeom which was a GC bug. Revert
not-so-efficient commit a94cab4e2b