rewrite ellisdivisible + wrong for p^2 when p-torsion [#2408]
e = ellinit([0, 0, 0, -25, 0])
q = ellmul(e, [-4,6], 4)
ellisdivisible(e,q,4) \\ = 0

It is valid to decompose the check for P = [n]Q one primary divisor of
n at a time, not one prime at a time which is wrong for p^2 in presence of
p-torsion. To reduce from p^k to the prime case, one could check whether
p | #E(K)_tor, and if so check recursively (e.g, depth first) the tree of
solutions. I.e. if P = [p]Q, then test whether one of all possible Q
(at most p^2) is of the form [p^{k-1}]Q', etc., each atomic
test "A = [p]B ?" involving an nfroots. This is likely to be more
expensive than a single nfroots with p^k. Esp. if there is large
p-torsion but small p^k-torsion because there will be few local solutions
to the latter test, even if the degree is p^{2k} instead of p^2 (repeatedly!).